Moving all authentication, token, user and role checks to base resource
