+++ resource_allocator/resources/request.py
@@ -9,8 +9,7 @@ from resource_allocator.schemas.request import (
     RequestRequestSchema,
     RequestResponseSchema,
 )
-from resource_allocator.resources.base import BaseResource, CRUDResource
-from resource_allocator.managers.user import auth, role_required
+from resource_allocator.resources.base import BaseResource, CRUDResource, auth, role_required
 
 
 class RequestResource(CRUDResource):
